requirements.txt
numpy
opencv-python
blis
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
pydantic
spacy
PyPDF2
python-docx
requests
pydantic[email]
jinja2
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pytesseract
pdfplumber
sqlalchemy
